Define a separate protocol for the runner used by Project
The Project class needs something to run scripts and programs, and up to
this point that has been the ScriptRunner class from
pytest-console-scripts. But that makes every single use of Project
dependent on the a script_runner fixture - unless we decide to start
constructing custom ScriptRunner instances, but that would be a little
awkward since that class expects to be given things like option values
for the console-scripts plugin. It would be much nicer if we had
the option to swap out the runner and potentially use something simpler.

This commit makes that change by defining a new protocol type which
describes exactly the behavior we need for a runner. That behavior is
provided by ScriptRunner.run() (or a simple wrapper around it, if using
an old version of pytest-console-scripts where the signature is
different), but now we can use other implementations as well.

In addition, this commit also defines a new protocol for the return type
of a runner. As with the runner protocol, the return type protocol is
compatible with the version from pytest-console-scripts but can also be
implemented with something custom.