Pass the runner into the Project.run() instead of Project()
This commit changes the Project class so that we provide a runner to
the run() or run_cli() method, rather than to the constructor. This
means we can now create a Project without a runner, and only provide one
as needed in tests which actually use that functionality. In particular,
the live distribution package tests have no more need to accept a runner
at all.