Add configuration for towncrier
This commit sets up the configuration for towncrier. We'll put changelog
fragments in the directory changelog.d, which is being created (or
technically a stub file to ensure it always exists), and they will be
accumulated into CHANGELOG.rst, which I think is a pretty common name
and is consistent with other changelog-related terminology. I've renamed
NEWS.rst to CHANGELOG.rst accordingly.