Add a pre-commit hook to check changelog fragments
This hook just does some basic sanity checks on the changelog fragments.
I've chosen not to use the update hook because I'm not sure how it would
interact with setuptools-scm (which changes the version number on every
commit), and it might not even be a good fit for our workflow anyway.