Allow running towncrier from tox in a way that works
I believe that jaraco.develop.finalize is not the right tool for us to
use when preparing for a release. For one thing, when I run it I get
errors about a missing dependency (tomlkit); probably more importantly,
it seems to do a few things that are incompatible with the way we have
this project set up, like creating an unused newsfragments directory and
adding some random files to Git. I haven't fully investigated this but I
think there is minimal benefit in using jaraco.develop.finalize compared
to just running towncrier directly.

So I've modified the tox environemnt to do just that. I changed it to
actually install the package to avoid and problems with missing
dependencies (although it's possible that wasn't the right way to fix it
but we can figure that out later) and I replaced the previous command
with a simple invocation of `towncrier build`. I also changed
the environment name to `towncrier` to reflect its new role.

Note that when preparing a release, we'll have to pass the --version
option to towncrier (or, to `tox run -e towncrier`) manually for now. I
hope that in the future we can automate that process.