Fix the non-running of slow tests in CI
When I added the feature to skip slow tests in CI under most conditions,
I made the classic mistake of using the logical ternary operator pattern
A && B || C with B being something that evaluates to boolean false. So
the result was always C, i.e. '-m "not slow"', which skips the slow
tests.

In this commit I'm flipping the sense of the condition so that I can put
the truthy value '-m "not slow"' in position B, which makes it work
properly.

I hate Javascript :-( or more precisely, the miniature Javascript-based
logic expressions that Github supports in workflow definitions, along
with the lack of strong typing.