Tell ReadTheDocs to ignore local changes to conf.py
When ReadTheDocs builds our documentation, it makes some local changes
to docs/conf.py. setuptools_scm then checks the status of the Git
working tree when determining the package version, and comes up with
a version number suitable for a local development version, like
"0.1.1.dev16+g124dc35.d20231012", which then appears in the generated
documentation. But we want a final release version number like 0.1.1,
at least for versions of the documentation generated from final release
tags.

The solution recommended in ReadTheDocs' documentation
  https://docs.readthedocs.io/en/stable/build-customization.html#avoid-having-a-dirty-git-index
is set the assume-unchanged bit on the modified working tree files so
Git doesn't treat them as modified. This commit implements that change.