Use a full clone when building on ReadTheDocs
Normally ReadTheDocs runs a shallow git clone command, which won't give
it access to the Git history needed to determine the correct version.
This commit adds a step that will change the clone to a full one, with
tags available. (The --tags option is redundant, I think, but I included
it to be explicit that tags are what we want.) This is a variant of
a recommendation from the ReadTheDocs documentation:
    https://docs.readthedocs.io/en/stable/build-customization.html#unshallow-git-clone