Move test-specific state out of DistributionPackage
This commit does a pretty major refactoring of the distribution package
test support code. The main change is to remove all test-specific state
and state-changing operations from the DistributionPackage class and put
them in DistributionPackagePreparation, so now the workflow is that we
use a list of DistributionPackage objects to parametrize the tests, and
in each test we "convert" the DistributionPackage to an instance of
DistributionPackagePreparation which is then used and mutated as needed
for the test. Without doing this, it wouldn't really make sense to use
DistributionPackage instances to parametrize tests because they would
change each time they were used.

A big part of that change involved creating a hierarchy of
DistributionPackagePreparation classes, because each type of
DistributionPackage needs its own DistributionPackagePreparation
subclass to store relevant state.

That change turned out to be a very complex one, and while working on it
I also made some other changes that turned out to be pretty difficult to
disentangle afterwards. So I'm leaving these additional changes as part
of this commit:

- Improved use of cached properties and function return values to avoid
  downloading or extracting the same files more than once
- Better organization of the code that computes core reference metadata
  so that now each possible metadata source (wheel file or metadata file
  from PyPI) is its own separate method