Add test support function for parsing a message into StandardMetadata
Different distribution packages store their metadata in different ways
which can be hard to compare directly, most notably due to different
versions of the core metadata standard. For example, the Requires and
Requires-Dist headers are equivalent-ish in the sense that they both
express dependencies, but technically any recent core metadata listing
such in a PKG-INFO file could use either one (or maybe even a mix of
both, not that I've seen that). This makes it difficult to compare
the expected metadata for a package, obtained from the package itself,
with the metadata generated from setuptools-pyproject-migration, since
they could adhere to different versions of the core metadata standard
and use different sets of headers in other ways.

After lots of false starts, I decided that the most practical way to
do that comparison is to parse the core metadata (PKG-INFO or whatever)
back into some canonical format, such as the StandardMetadata object
from pyproject-metadata. But pyproject-metadata doesn't allow parsing
its RFC822Message class back into a StandardMetadata, it only goes
the other way. So I wrote us a function to do that conversion, which is
what I'm adding here.

I haven't extensively tested this function, and it could very well have
bugs - handling all the different versions of the core metadata standard
is a little complex. But it seems to work well enough for the moment. We
can always add more tests (ah, the irony of having to test the test
support code) and fix issues that may show up in the future.