Add make_importable flag
In this commit I'm introducing a new flag that will cause the root
directory of a distribution package project to be added to sys.path
while that package is being tested. Some packages expect their own
code to be available when running their setup.py, and this accounts
for that requirement.