Add a new console script name and deprecate the old one
In this commit I'm changing the primary name of the console script to
match the name of the package, setuptools-pyproject-migration. This will
allow people to run the program directly with
  pipx run setuptools-pyproject-migration
without an explicit installation step.

We could keep both console script names, but I think that's going to be
more confusion than it's worth, so I've marked the old name
setup-to-pyproject as deprecated. This was simply done by writing
a wrapper function that issues a FutureWarning and redirecting the old
console script name to call it. (Our users don't need anything fancy.)