Fix Markdown syntax in documentation
This snippet of documentation was in an .rst file and should have been
in Restructured Text syntax, but it was actually in Markdown, so it
wasn't rendering properly in the final documentation. I'm fixing that
with this commit.

It would be nice if we had a pre-commit hook to check .rst syntax, it
probably would have caught this, but I haven't found a really good hook
for that yet.