Fix link syntax in docstring
I used the wrong link syntax in a docstring, so I'm fixing that here.

This is another case where a suitable pre-commit hook would have helped,
though in this case we'd need one able to check RST syntax in docstrings
specifically. I'm not sure if there is a good hook that does that.