Require setuptools 66.1 for Python 3.12+
Older versions of setuptools fail to import distutils on Python 3.12,
presumably due to the removal of distutils from the standard library.
This bug was fixed in https://github.com/pypa/setuptools/pull/3685, and
setuptools 66.1 was the first version released that includes the fix.
(I'm not sure exactly how that PR fixed the bug, but it doesn't matter.)

Because setuptools 66.0 understandably does not declare itself to be
incompatible with Python 3.12, I'm recording that constraint here.