Update changelog
I went through the process of running towncrier locally to generate
a changelog fragment, then ran it again to add that fragment into
the changelog, and manually combined it with the previous version.

This normally shouldn't be necessary, I'm only doing it this time to
help recover from a failed release.