Conditionally disable the pytest distribution package test
When running tests under Python 3.12 with setuptools_scm==3.4.1, we run
into a problem where a call to setuptools.setup() tries to load a newer
version of setuptools_scm, and that conflicts with the older version
that is already in sys.modules. I haven't fully figured out the details,
but the gist is that setuptools attempts to read and install the entry
point specifications from the newer version of setuptools_scm while
using the older version's code, and there is one particular entry point
whose value refers to the setuptools_scm._integration module which
doesn't exist in the older version's code. This winds up breaking every
test that runs after the older version of setuptools_scm gets loaded.

We have various options for fixing this, but I don't want to hold up
the release any longer while figuring out what the best way to do it is.
(I'm guessing that the best approach is just to use virtual environments
for distribution package tests, but I want to explore the options.) So
I'm disabling the distribution package test which triggers this error
when using Python>=3.12 and setuptools_scm<6. We can implement a proper
fix later.