Reduce the pre-commit autoupdate frequency to monthly
It's kind of annoying to deal with weekly updates because we always have
to revert the changes to the ruff hook, at least as long as we want to
keep supporting Python 3.6. So I'm changing it to monthly. We could
decide to further reduce the frequency or drop usage of pre-commit.ci
entirely in the future.

Closes #109