Allow for missing long_description key in command options metadata
The command_options attribute of a Distribution contains the CLI options
passed to that command, but it also contains project metadata specified
in setup.cfg, under the "metadata" key. So a project might have metadata
in this value that doesn't have the long_description key, if
the long_description argument was passed to setup.py instead.
Accordingly, we need to check both whether the "metadata" key exists,
and whether the "long_description" key exists within it, to avoid
getting an error. This commit implements that change.