Refactor the code for handling long descriptions
In upcoming commits I'm going to be changing the code that handles
the long description/readme fields in a pretty significant way, and that
code is going to get rather complex, so I decided to split it out into
a separate module to aid readability. I also converted it from a simple
block to a class and some helper functions. The class stores the text,
content type, and filesystem path of the readme element, and it provides
some validation of those values before they get serialized into the data
structure that will become pyproject.toml.

This commit doesn't include any change in functionality, it's just
refactoring the code so that future functional changes will be easier to
understand.