Allow writing readme info without a filename
It turns out pyproject.toml does support putting the content of
the readme attribute directly inline, so this commit changes the logic
to do that when no filename can be determined. Since there are many
cases where setup.py reads the content of a file and gives that content
to the long_description argument of setup(), the code will look for
a file containing the readme text and use that if found.