Add a command-line option to override the content type
pyproject.toml supports giving readme information in any of three forms:

- A filename with a content type
- A filename without a content type, but it's expected that the type
  will be inferred from the file's extension
- A raw string with a content type

setuptools, however, also allows giving the long description as a raw
string with no associated content type. In this case it's ambiguous what
the corresponding data structure in pyproject.toml should be.

This commit adds an option for the user to specify the content type
explicitly and resolve the ambiguity. I chose this approach based in
part on a Mastodon poll:
  https://techhub.social/@diazona/112736102428118730
The leading option in the poll was to make the user give an explicit
content type, but on the other hand most of the people who replied
suggested that it'd be better to guess, falling back to text/plain if no
other content type should be determined. We could certainly take either
approach, but ultimately I was more swayed by the responses. Plus, the
program already tries to guess a readme file (although admittedly that's
a more educated guess than for the content type), so we kind of have
a precedent for making guesses to fill in unknown values rather than
failing and forcing the user to provide them explicitly.