Implement a marker that distributes other markers across tests by name
In this commit I'm adding a new marker, `distribute`, which takes a dict
mapping test names to "sub-markers" and will apply each "sub-marker" to
any test with the corresponding name (without parameters).

Normally this isn't useful because we could just apply the sub-markers
to their respective tests directly. But when we have a test class which
is also parametrized, as with TestExternalProject, it sets up a "matrix"
of parameters and test methods, and without something like this feature,
there's no good way to apply a marker to a specific combination of test
method and parameter. We need it to indicate that specific test methods
are expected to fail (xfail) only for specific parameter values.

As an example of the benefit of this marker, once I started using it I
realized that test_readme for pytest-localserver was actually failing,
even though it was supposed to be fixed by the previous merge (commit
f5a110fcc2d82076fd18de82d2131404d29e7ee9). That was previously masked
because the xfail marker had been blanket applied to all tests that use
pytest-localserver as the external project.