Add issue links to xfail markers
With the new `distribute` marker that allows us to apply markers to
specific combinations of test method and parameter value, I can now add
reasons to each of the xfailing tests. In this commit I'm using that to
indicate which GitHub issue corresponds to each expected failure, for
better tracking.