Exclude conftest.py from type checking
I'm going to be adding another conftest.py file in the next commit, and
when that happens, mypy complains about conflicting names. As
a quick-and-dirty resolution to this issue, I'm excluding conftest.py
from type checking. Hopefully in the future we can find a better way to
handle this issue that allows the conftest.py files to be type-checked
as well.