Add testing with Python 3.13
The final beta of Python 3.13 is now available, so I'm adding it to
the list of versions we test. It's not required to pass for now since
3.13 is not final.

In order to get more exposure with Python 3.13 testing, I chose to
enable it in pre-merge tests, not only releases. It won't block merges
if something breaks since the test is not required to pass.