Add coverage.py configuration to pyproject.toml
This configuration enables branch coverage and specifies the packages
for which coverage should be measured. Previously, there was nothing
telling coverage.py not to measure coverage for files generated during
testing, so it would pollute the output with meaningless coverage
numbers for a bunch of temporary files.

I chose to put the configuration in pyproject.toml rather than any of
the other configuration files coverage.py supports because TOML is
a better specified format than INI, so I prefer to use it when possible.