Explicitly specify coverage configuration file
The documentation of pytest-cov advises that because of how coverage.py
handles config file lookup, when using a file other than .coveragerc to
configure it, it's important to specify that file name explicitly. (I
suspect it has to do with pytest-cov invoking coverage.py in a way that
overrides its normal configuration lookup mechanism, but I don't know
the details.) Accordingly, I'm explicitly setting the coverage.py config
file to pyproject.toml.

This does mean that our tox configuration doesn't support running tests
without coverage, but we've never supported that. If necessary we can
create non-coverage testenvs in a future version, it's not hard.