Split keywords in test support code
Even though the core metadata allows multiple Keywords headers, in
practice all a package's keywords seem to be combined into one header
joined by commas, or maybe spaces for some older packages. This commit
changes the test support code that parses core metadata to account for
that by splitting the Keywords header (if single-valued). The delimiter
is chosen based on the metadata version: in core metadata version 1.x,
it allows spaces as a separator, but in version 2.x, it only allows
commas.