Run slow tests in pre-/post-merge testing
I realized that we were not running the distribution package tests in
pre-merge tests (on pull requests) or post-merge tests (after merging to
main), and that meant there were failures in the distribution package
tests which were not getting caught. This commit fixes that so that slow
tests will run during pre- and post-merge testing.

Our test suite doesn't really take long enough to make it important to
skip slow tests anyway.