Use an empty list to represent no keywords
In commit 09e4dfc665c805b41154e026d5612e217fa2adfa I refactored the code
in our test support package that handles keywords, and I accidentally
changed it to use a list with a single empty string as the default value
if there are no keywords in the package, when it should be using
an empty list instead. In this commit I'm fixing that error. This allows
our distribution package tests to pass again.