Replace None email address in StandardMetadata with empty string
pyproject-metadata's StandardMetadata.from_pyproject() method has
an apparent bug where, if it is handling author or maintainer metadata
that specifies a name but no email address, it uses None to represent
the email address even though the declared type of the field is str.
I reported this upstream at
  https://github.com/pypa/pyproject-metadata/issues/126
but there's been no activity on the report so far. So, until that
changes, I'm adding code to work around that bug by replacing None in
the email address field with an empty string.