Handle "optional" dependencies with no extra
Versions of setuptools prior to 68.2.0 have an issue where they will
misrepresent dependencies that have a constraint but no extra as
optional dependencies with an empty extra string. For example, if
a setup.cfg file includes this:

  [options]
  install_requires =
    importlib-metadata>=0.12;python_version<"3.8"

that dependency should be parsed as a required dependency on all Python
versions less than 3.8, but these old versions of setuptools parse it as
if it were this instead:

  [options.extras_require]
  =
    importlib-metadata>=0.12;python_version<"3.8"

(with an empty key under options.extras_require). This causes test
failures on older versions of Python or anywhere else we have an old
version of setuptools to work with.

To address this, in this commit I'm changing the logic that parses
dependencies and optional dependencies so it adds anything associated
with an empty extra to the set of required dependencies. I also added
some logging statements that were helpful in diagnosing this error.