Add a test of a required dependency with a constraint
This test checks whether the program properly works around a bug in
setuptools<68.2 where required dependencies with constraints get treated
as if they listed as optional dependencies under an extra with an empty
name.

Previously, the only test that would have exposed this issue was the
pytest distribution package test. In pre- and post-merge testing, we
were only running that test with old versions of setuptools on Python
3.12, and under those particular circumstances the pytest distribution
package test would be skipped, so the issue never showed up until we
made a release. This test would allow us to catch it normally.