Enable testing with minimum dependencies on Python 3.8
Previously, in pre- and post-merge testing, we would only test with
the oldest allowed versions of our dependencies on Python 3.12. However,
the pytest distribution package test gets skipped on Python 3.12, which
means we're sacrificing an opportunity to detect issues that arise with
pytest when using old versions of setuptools and friends.

This commit also enables minimum dependency testing on Python 3.8, so we
have a chance to find those errors before making a release.