Skip tests causing setuptools-scm conflict on all Python versions
We've had a long-standing issue on Python 3.12 where running our
distribution package test suite on pytest would mess up the whole test
session, due to conflicting versions of setuptools-scm being installed
in the same environment. Well, it turns out this issue occurs on other
versions of Python as well, not only 3.12. So I'm changing the condition
in the code that we use to skip the pytest distribution package tests so
that it will be true on any version of Python, not only 3.12+. This
prevents test failures in CI.