Use a different implementation of sdist and wheel filename matching
While adding the extrainterpreters package as a test case, I discovered
a bug where the regex we use for matching sdist and wheel filenames is
not correctly parsing certain prerelease versions. In this commit I'm
fixing that by outsourcing the parsing of wheel filenames to the package
wheel_filename, and writing a more accurate regex to parse sdist
filenames. I've added a tuple representing the components of an sdist
filename to parallel the structure that wheel_filename uses.

The new dependency is only needed in testing so it's relatively low cost
to add it.