Add tests of filename matching
This commit adds some tests of the sdist/wheel filename matching
function added in the previous commit. It's not terribly valuable to
have a lot of tests of wheel filename matching because it's being
handled by a separate package which (presumably) has its own tests,
but it'll be good to have tests regardless in case we ever switch to
a different implementation, and it costs us very little to add them.