test_support metadata: Add `mypy` annotation to silence warning
Icky, and really not advised, but this feels like a `mypy` bug.
`StandardProject` is expecting a `str | None`, which is Python 3.10
syntax for `typing.Union[str, None]` (aka `typing.Optional[str]`).

This *should* have been fine; we're assigning a `List[Tuple[str, str]]`
so the `Optional[str]` is always assigned a `str` -- this is fine.
`mypy` though thinks this isn't valid.

Seems to be a similar issue to this reported issue:
https://github.com/python/mypy/issues/14602