setup.cfg: Bump `setuptools-metadata` requirements
The older version required `List[Tuple[str, str]]` for `maintainers` and
`authors`, the newer one requires `List[Tuple[str, Optional[str]]]`,
`mypy` won't accept the former as a substitute for the latter.

`setuptools-metadata` 0.8.1 made this change, then 0.9.0 further broke
the API established in pre-0.8.1 releases.  So to avoid a lot of test
complexity, we'll just settle for 0.9.0 going forward.