Update Python versions in GHA workflows to drop 3.6
We are no longer able to test on Python 3.6 using the GitHub-provided
standard actions, so I'm removing that version from our workflows. In
addition to dropping Python 3.6, this updates the other Python versions
we test on to reflect the edge cases we want to test, such as the latest
stable version (currently 3.13) and the oldest non-EOL version (3.9).