Allow manual release testing
In this commit I'm adding a workflow_dispatch trigger to the release
workflow, which means at any time we can manually invoke the workflow to
run a full suite of pre-release tests. I added conditions to the actual
publish jobs so that manually invoked workflows will not attempt to
upload to PyPI or Test PyPI, but they will do everything else. I also
added the run-name entry in the YAML file to put a more descriptive name
on manually triggered runs, so we don't get confused about which runs
are actually publishing packages.