GH Issue 77: Add `spdx-lookup` module to dependencies
It is unmaintained, so not a good long-term solution, but it does have
the nice feature that it can match license text to a probable license.

So in addition to _our_ internal parsing logic, which will do the heavy
lifting of actually validating SPDX identifiers, this may help migrate
projects over to the new format.