Dynamically determine Python prerelease versions in CI
When we run tests in CI, the testing job is supposed to fail if
the tests fail, but only for stable versions of Python, not for
prereleases. We use a continue-on-error setting in the workflow file to
implement this. When Python 3.13 was released, I guess we forgot to
update the setting to make Python 3.14 the first version that could fail
tests without failing the job, so for a long time our CI testing could
have failed on Python 3.13 and we probably wouldn't have noticed if it
didn't also fail on other versions.

To fix that, in this commit I'm changing the test workflow to
dynamically determine whether the active Python version is a stable or
prerelease version, and use that to decide whether to allow tests to
fail. In order to make it work, I had to change from a continue-on-error
setting on the whole job to continue-on-error on only the specific steps
that run the tests, but that's better anyway because it means CI will
now catch failures that happen during setup even on prerelease Python
versions.