Require a changelog fragment in each pull request
It's easy to forget to include changelog fragments, so in this commit
I'm adding a job that will fail on any pull request that doesn't add
a file under changelog.d/, unless it looks like it's regenerating
the changelog (removing all fragments from changelog.d and updating
CHANGELOG.rst).

I tested the script by extracting it from the YAML file and piping to
bash, with various combinations of the environment variables set.