Stop using a real filename for generated code in the console script
If our console script, setuptools-pyproject-migration, is run on
a project that has setup.cfg but not setup.py, it uses a compile() call
to simulate a bare-bones setup.py file, allowing us to essentially run
a script that doesn't exist. However, when the console script is run in
testing with coverage enabled, that causes a warning because coverage.py
looks at the name given to compile() and expects to find a corresponding
file on disk - but it doesn't exist.

To fix that issue and remove the warning, I'm switching the filename
given to compile() to surround it with <>, as is typical when compiling
code that didn't come from a real file. coverage.py knows to recognize
angle brackets around a filename and not look for a corresponding source
file.