GH77: core: Use `spdx-lookup` to match license text to a license.
`spdx-lookup` is not maintained, which is definitely a big downside as
we run the risk of bitrot, and doesn't implement the full SPDX
expression parsing features, but it has one redeeming feature: it can
look at a license text, and identify it with a confidence measurement,
as being a particular license.

This means we can suggest to the user (with a warning) what the most
likely matching SPDX license is for a given text.  If we don't get a
match, we bail here with a `ValueError`, but if there's a match, we
issue a warning to get the user to check it.